+++ R/calc_genoprob_gbs.R
@@ -1,0 +1,13 @@
+# calc_genoprob_gbs
+
+calc_genoprob_gbs <-
+    function(counts, map, error_prob1=0.002, error_prob2=0.002,
+             map_function)
+{
+
+
+
+
+
+
+}