+++ NEWS.md
@@ -1,3 +1,9 @@
+## xoiGBS 0.1.4 (2025-11-05)
+
+- Added calc_genoprob_gbs to calculate genotype probabilities from
+  genotyping-by-sequencing (GBS) data (counts of alleles at SNPs) for
+  a backcross.
+
 ## xoiGBS 0.1.3 (2025-03-05)
 
 - A new package, to analyze recombination and crossover interference